+++ examples/example_advertiser_report_installs.py
@@ -32,7 +32,7 @@
 #  @author    Jeff Tanner <jefft@tune.com>
 #  @copyright 2014 TUNE, Inc. (http://www.tune.com)
 #  @license   http://opensource.org/licenses/MIT The MIT License (MIT)
-#  @version   $Date: 2014-12-19 15:59:09 $
+#  @version   $Date: 2014-12-21 13:25:20 $
 #  @link      https://developers.mobileapptracking.com/tune-reporting-sdks @endlink
 #
 # You can use the Logs report in the same way as the Actuals reports, but
@@ -108,18 +108,18 @@ class ExampleAdvertiserReportInstalls(object):
             advertiser_report = AdvertiserReportInstalls()
 
             print("")
-            print("======================================================")
-            print(" Fields of Advertiser Report Installs records.        ")
-            print("======================================================")
+            print("===========================================================")
+            print(" Fields of Advertiser Report Installs records.             ")
+            print("===========================================================")
 
             response = advertiser_report.fields(TUNE_FIELDS_RECOMMENDED)
             for field in response:
                 print(str(field))
 
             print("")
-            print("======================================================")
-            print(" Count Advertiser Report Installs records.            ")
-            print("======================================================")
+            print("===========================================================")
+            print(" Count Advertiser Report Installs records.                 ")
+            print("===========================================================")
 
             response = advertiser_report.count(
                 start_date,
@@ -138,9 +138,9 @@ class ExampleAdvertiserReportInstalls(object):
             print(str(response.data))
 
             print("")
-            print("======================================================")
-            print(" Find Advertiser Report Installs records.                 ")
-            print("======================================================")
+            print("===========================================================")
+            print(" Find Advertiser Report Installs records.                  ")
+            print("===========================================================")
 
             response = advertiser_report.find(
                 start_date,
@@ -160,9 +160,9 @@ class ExampleAdvertiserReportInstalls(object):
             print(str(response))
 
             print("")
-            print("==========================================================")
-            print(" Export Advertiser Report Installs CSV report             ")
-            print("==========================================================")
+            print("===========================================================")
+            print(" Export Advertiser Report Installs CSV                     ")
+            print("===========================================================")
 
             response = advertiser_report.export(
                 start_date,
@@ -184,9 +184,9 @@ class ExampleAdvertiserReportInstalls(object):
             print(" CSV Job ID: {}".format(job_id))
 
             print("")
-            print("=================================================================")
-            print(" Fetching Advertiser Report Installs CSV report                    ")
-            print("=================================================================")
+            print("===========================================================")
+            print(" Fetching Advertiser Report Installs CSV                   ")
+            print("===========================================================")
 
             export_fetch_response = advertiser_report.fetch(
                 job_id,
@@ -199,9 +199,9 @@ class ExampleAdvertiserReportInstalls(object):
             print(" CVS Report URL: {}".format(csv_report_url))
 
             print("")
-            print("========================================================")
-            print(" Read Advertiser Report Installs CSV report     ")
-            print("========================================================")
+            print("===========================================================")
+            print(" Read Advertiser Report Installs CSV                       ")
+            print("===========================================================")
 
             csv_report_reader = ReportReaderCSV(csv_report_url)
             csv_report_reader.read()
@@ -209,7 +209,7 @@ class ExampleAdvertiserReportInstalls(object):
 
             print("")
             print("===========================================================")
-            print(" Export Advertiser Report Clicks JSON report ")
+            print(" Export Advertiser Report Clicks JSON                      ")
             print("===========================================================")
 
             response = advertiser_report.export(
@@ -232,9 +232,9 @@ class ExampleAdvertiserReportInstalls(object):
             print(" JSON Job ID: {}".format(job_id))
 
             print("")
-            print("========================================================")
-            print(" Fetching Advertiser Report Installs JSON report.         ")
-            print("========================================================")
+            print("===========================================================")
+            print(" Fetching Advertiser Report Installs JSON                  ")
+            print("===========================================================")
 
             export_fetch_response = advertiser_report.fetch(
                 job_id,
@@ -254,9 +254,9 @@ class ExampleAdvertiserReportInstalls(object):
             print(" JSON Report URL: {}".format(json_report_url))
 
             print("")
-            print("========================================================")
-            print(" Read Advertiser Report Installs JSON report    ")
-            print("========================================================")
+            print("===========================================================")
+            print(" Read Advertiser Report Installs JSON                      ")
+            print("===========================================================")
 
             json_report_reader = ReportReaderJSON(json_report_url)
             json_report_reader.read()