+++ tune_reporting/api/advertiser_report_clicks.py
@@ -36,12 +36,12 @@ TUNE Management API '/advertiser/stats/clicks/'
 #  @author    Jeff Tanner <jefft@tune.com>
 #  @copyright 2014 TUNE, Inc. (http://www.tune.com)
 #  @license   http://opensource.org/licenses/MIT The MIT License (MIT)
-#  @version   $Date: 2014-12-19 11:19:31 $
+#  @version   $Date: 2014-12-21 13:25:20 $
 #  @link      https://developers.mobileapptracking.com/tune-reporting-sdks @endlink
 #
 
 from tune_reporting.base import (
-    ReportsLogsEndpointBase
+    AdvertiserReportLogsBase
 )
 
 
@@ -49,7 +49,7 @@ from tune_reporting.base import (
 #
 #  @example example_reports_clicks.py
 #
-class AdvertiserReportClicks(ReportsLogsEndpointBase):
+class AdvertiserReportClicks(AdvertiserReportLogsBase):
     """TUNE Management API endpoint '/advertiser/stats/clicks'
     """
 
@@ -58,7 +58,7 @@ class AdvertiserReportClicks(ReportsLogsEndpointBase):
     def __init__(self):
         """The constructor.
         """
-        ReportsLogsEndpointBase.__init__(
+        AdvertiserReportLogsBase.__init__(
             self,
             "advertiser/stats/clicks",
             True,